Refactor: File Package
Related to: #21
