+++ src/test/java/com/maruhxn/lossion/util/IntegrationTestSupport.java
@@ -1,7 +1,7 @@
 package com.maruhxn.lossion.util;
 
 import com.maruhxn.lossion.infra.EmailService;
-import com.maruhxn.lossion.infra.FileService;
+import com.maruhxn.lossion.infra.file.FileService;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ActiveProfiles;
